+++ srcs/structures/engine/map_params/utils/parsing/map/map_parsing.c
@@ -17,12 +17,12 @@ static t_bool   is_only_map_entries(char **map)
     t_vector    index;
 
     index = create_vector(0, 0);
-    while (map[index.y] != NULL)
+    while (map[(int)index.y] != NULL)
     {
         index.x = 0;
-        while (map[index.y][index.x] != '\0')
+        while (map[(int)index.y][(int)index.x] != '\0')
         {
-            if (is_valid_map_entry(map[index.y][index.x]) == false)
+            if (is_valid_map_entry(map[(int)index.y][(int)index.x]) == false)
                 return (false);
             index.x++;
         }
@@ -38,12 +38,12 @@ static t_bool   is_valid_spawn_point(char **map)
 
     index = create_vector(0, 0);
     flag = 0;
-    while (map[index.y] != NULL)
+    while (map[(int)index.y] != NULL)
     {
         index.x = 0;
-        while (map[index.y][index.x] != '\0')
+        while (map[(int)index.y][(int)index.x] != '\0')
         {
-            if (is_cardinal_point(map[index.y][index.x]) == true)
+            if (is_cardinal_point(map[(int)index.y][(int)index.x]) == true)
             {
                 if (flag == 0)
                     flag = 1;
@@ -66,12 +66,12 @@ static t_bool   is_valid_map(char **map)
     index = create_vector(0, 0);
     if (is_only_map_entries(map) == false)
         return (false);
-    while (map[index.y] != NULL)
+    while (map[(int)index.y] != NULL)
     {
         index.x = 0;
-        while (map[index.y][index.x] != '\0')
+        while (map[(int)index.y][(int)index.x] != '\0')
         {
-            if (is_tile(map[index.y][index.x]) == true && is_walled(map, index) == false)
+            if (is_tile(map[(int)index.y][(int)index.x]) == true && is_walled(map, index) == false)
                 return (false);
             index.x++;
         }
